# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Uno
  version: "1"
  description: |
    Uno's HTTP surface: the builder API under `/api`, which the SPA calls, and
    the public Stremio-protocol addon under `/u/{token}`, which Nuvio calls.
    Everything else on the server is the embedded SPA and is not described
    here.

    **Errors.** Every error under `/api`, an unknown path included, is JSON
    (`Error`): `error` holds one line of words and, for an error the builder
    acts on by kind, `code` names it (`ProfileNotFound`). Push is the one
    exception: past auth, profile resolution and body decoding it answers
    `PushResult` whether it succeeds or fails. The addon under `/u/{token}`
    answers errors in plain text (`text/plain; charset=utf-8`), which its
    clients don't read.

    **Request bodies.** Every body is decoded strictly: a field the request
    doesn't have, at any depth, is a 400 naming it (`invalid request body:
    unknown field "name"`). The default limit is 1 MiB, past which the answer
    is 413.

    **Auth.** Every `/api` operation except `GET /api/health` and
    `GET /api/config` takes a Nuvio access token as `Authorization: Bearer`.
    Any of them can answer 401 (missing, malformed, invalid or expired
    token), 403 (an account the server's allowlist doesn't list) or 502
    (`auth service unavailable`: Nuvio's signing keys couldn't be fetched).
    Any operation can answer 500 for a failure of Uno's own.

...

...
    TMDB key or TMDB refuses it.

...

...
servers:
...
security:
  - nuvioBearer: []
paths:
...
  /api/profiles:
    get:
      tags: [profiles]
      operationId: listProfiles
      summary: The caller's Nuvio profiles
      description: Nuvio's profile list, each with the picture Nuvio's apps show for it.
      responses:
        "200":
          description: Every profile on the caller's Nuvio account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PickerProfile"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/NuvioUnavailable" }
  /api/profiles/select:
    post:
      tags: [profiles]
      operationId: selectProfile
      summary: Pick a Nuvio profile to build for
      description: |
        Matches `profile_index` against the caller's live Nuvio profiles and
        resolves the Uno profile for that slot, creating it on first pick.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile_index]
              properties:
                profile_index:
                  type: integer
      responses:
        "200":
          description: The Uno profile, with its addon manifest URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectedProfile"
        "400":
          description: The body doesn't decode, or the caller's Nuvio account has no profile at that index.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "502": { $ref: "#/components/responses/NuvioUnavailable" }
  /api/account/tmdb-key:
    get:
      tags: [account]
      operationId: getTMDBKey
      summary: Whether the caller has saved a TMDB key
      responses:
        "200":
          description: The key's status; the key itself is never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TMDBKeyStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SharedKeyMode" }
    put:
      tags: [account]
      operationId: putTMDBKey
      summary: Save the caller's TMDB key
      description: |
        Saves the key, replacing any saved one, once it has the shape of a
        TMDB API key and TMDB accepts it. Nothing is saved otherwise.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key:
                  type: string
                  description: A TMDB API Key (32 characters, 0–9 and a–f), not a Read Access Token.
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TMDBKeyStatus"
        "400":
          description: The key isn't shaped like a TMDB API key, or TMDB refused it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SharedKeyMode" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
    delete:
      tags: [account]
      operationId: deleteTMDBKey
      summary: Remove the caller's TMDB key
      responses:
        "204":
          description: Removed, or there was none.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SharedKeyMode" }
  /api/catalogs/preview:
    post:
      tags: [recipes]
      operationId: previewCatalog
      summary: Run a recipe against TMDB
      description: |
        Returns the first page of titles a recipe finds, saving nothing. The
        provider is always `tmdb`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, params]
              properties:
                type:
                  $ref: "#/components/schemas/CatalogType"
                params:
                  $ref: "#/components/schemas/RecipeParamsString"
                genre:
                  type: string
                  description: A genre name from the recipe's genre options, narrowing the preview.
      responses:
        "200":
          description: The first page of titles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preview"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/catalogs/genre-options:
    post:
      tags: [recipes]
      operationId: catalogGenreOptions
      summary: The genres a recipe can be narrowed by
      description: The same list the manifest advertises as the catalog's genre extra.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, params]
              properties:
                type:
                  $ref: "#/components/schemas/CatalogType"
                params:
                  $ref: "#/components/schemas/RecipeParamsString"
      responses:
        "200":
          description: The genre options.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Genre"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/p/{profileIndex}/library:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
    get:
      tags: [catalogs]
      operationId: getLibrary
      summary: The profile's own catalogs and collections, and what a push would change
      description: |
        Everything the builder shows for a profile in one read: its listed
        catalogs (none scoped to a collection), its collections with their
        folders and catalogs, and the rows a push of the stored Home would
        change in Nuvio. Unpushed Home edits live in the builder and are not
        in it. It is not paged: a profile holds at most 200 listed catalogs and
        50 collections, and a write that would add past either is a `400`.
      responses:
        "200":
          description: The library.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
        "400": { $ref: "#/components/responses/BadProfileIndex" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/ProfileNotFound" }
  /api/p/{profileIndex}/catalogs:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
    post:
      tags: [catalogs]
      operationId: createUserCatalog
      summary: Create a catalog
      description: |
        Creates a listed catalog. A catalog scoped to a collection is created
        only by that collection's save, as a folder's `new` entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogForm"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/ProfileNotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/p/{profileIndex}/catalogs/{catalogID}:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CatalogID"
    put:
      tags: [catalogs]
      operationId: updateUserCatalog
      summary: Save a catalog
      description: |
        Rewrites a listed catalog. A catalog scoped to a collection is saved
        through its collection's `catalog_edits` instead. An update never
        changes a catalog's scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogForm"
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
    delete:
      tags: [catalogs]
      operationId: deleteUserCatalog
      summary: Delete a catalog
      description: Allowed on a catalog on Home; Nuvio keeps its row until the next push.
      responses:
        "204":
          description: Deleted.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/catalogs/{catalogID}/duplicate:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CatalogID"
    post:
      tags: [catalogs]
      operationId: duplicateUserCatalog
      summary: Duplicate one of the profile's own listed catalogs
      description: |
        Writes a listed copy named "<name> (copy)", the name cut short to fit
        200 characters. The stored recipe is checked by the form validators
        only, with no TMDB call.
      responses:
        "201":
          description: The copy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/catalogs/{catalogID}/publish:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CatalogID"
    post:
      tags: [publishing]
      operationId: publishCatalog
      summary: Publish a listed catalog, or publish its update
      responses:
        "200":
          description: Published; the catalog with its publication.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/catalogs/{catalogID}/unpublish:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CatalogID"
    post:
      tags: [publishing]
      operationId: unpublishCatalog
      summary: Unpublish a catalog
      description: |
        Deletes its publication. One-way: subscribers keep their copies as
        their own, and publishing again starts a new publication.
      responses:
        "200":
          description: Unpublished; the catalog without a publication.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/catalogs/{catalogID}/changes-since-publish:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CatalogID"
    get:
      tags: [publishing]
      operationId: catalogChangesSincePublish
      summary: What publishing the catalog again would change
      description: The saved catalog against its last snapshot; `[]` if it was never published.
      responses:
        "200":
          description: The changes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnapshotChange"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/catalogs/{catalogID}/acknowledge-release:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CatalogID"
    post:
      tags: [community]
      operationId: acknowledgeReleasedCatalog
      summary: Acknowledge a released catalog
      description: |
        Clears the catalog's release mark, set or not, once the
        builder has told its owner the publication ended. Saving never clears it.
      responses:
        "200":
          description: Acknowledged; the released copies still unacknowledged.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReleasedCopy"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/collections:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
    post:
      tags: [collections]
      operationId: createUserCollection
      summary: Create a collection
      description: Every inline new catalog's recipe is checked against TMDB first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionForm"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/ProfileNotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/p/{profileIndex}/collections/{collectionID}:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CollectionID"
    put:
      tags: [collections]
      operationId: updateUserCollection
      summary: Save a collection
      description: |
        Rewrites the collection, its folders and, through `catalog_edits`, its
        scoped catalogs in one transaction. A subscribed copy can't be saved
        (400). Every inline recipe is checked against TMDB first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionForm"
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
    delete:
      tags: [collections]
      operationId: deleteUserCollection
      summary: Delete a collection
      description: Allowed on a collection on Home; Nuvio keeps it until the next push.
      responses:
        "204":
          description: Deleted.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/collections/{collectionID}/duplicate:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CollectionID"
    post:
      tags: [collections]
      operationId: duplicateUserCollection
      summary: Duplicate one of the profile's own collections
      description: |
        Copies the whole tree, scoped catalogs included, in one transaction.
        The title gets a " (copy)" suffix, cut short to fit 200 characters.
      responses:
        "201":
          description: The copy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/collections/{collectionID}/publish:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CollectionID"
    post:
      tags: [publishing]
      operationId: publishCollection
      summary: Publish a collection, or publish its update
      description: |
        A subscribed copy is refused; a collection using catalogs added from
        Community is not.
      responses:
        "200":
          description: Published; the collection with its publication.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/collections/{collectionID}/unpublish:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CollectionID"
    post:
      tags: [publishing]
      operationId: unpublishCollection
      summary: Unpublish a collection
      description: |
        Deletes its publication. One-way: subscribers keep their copies as
        their own, and publishing again starts a new publication.
      responses:
        "200":
          description: Unpublished; the collection without a publication.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/collections/{collectionID}/changes-since-publish:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CollectionID"
    get:
      tags: [publishing]
      operationId: collectionChangesSincePublish
      summary: What publishing the collection again would change
      description: |
        The saved collection tree, its catalogs included, against its last
        snapshot; `[]` if it was never published.
      responses:
        "200":
          description: The changes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnapshotChange"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/p/{profileIndex}/collections/{collectionID}/acknowledge-release:
    parameters:
      - $ref: "#/components/parameters/ProfileIndex"
      - $ref: "#/components/parameters/CollectionID"
    post:
      tags: [community]
      operationId: acknowledgeReleasedCollection
      summary: Acknowledge a released collection
      description: |
        Clears the collection's release mark, set or not, once the
        builder has told its owner the publication ended. Saving never clears it.
      responses:
        "200":
          description: Acknowledged; the released copies still unacknowledged.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReleasedCopy"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    nuvioBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Nuvio access token, verified against Nuvio's signing keys.
...
  responses:
...
    Unauthorized:
      description: Missing, malformed, invalid or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    CatalogType:
      type: string
      enum: [movie, series]
...